@@ -23,7 +23,7 @@
 		"eslint-config-standard": "^17.1.0",
 		"obsidian": "latest",
 		"obsidian-typings": "^2.0.0",
-		"svelte": "^4.2.19",
+		"svelte": "^5.43.6",
 		"svelte-preprocess": "^6.0.2",
 		"tslib": "2.4.0",
 		"typescript": "^5.6.2"
